javascriptのリファクタリング
