+++ app/controllers/api/comments_controller.rb
@@ -1,4 +1,4 @@
-class CommentsController < ApplicationController
+class Api::CommentsController < Api::ApplicationController
   before_action :authenticate_user!, only: [:create]
 
   def index
